increases safety and reduces complexity in ast evaluator
